+++ src/Jieba/PosTagConstant.php
@@ -1,0 +1,130 @@
+<?php
+
+namespace Jieba;
+
+/**
+ * Class PosTagConstant. This file is generated with script /bin/genPosTagConstant.php.
+ *
+ * @package Jieba
+ * @see https://github.com/deminy/jieba-php/blob/master/bin/genPosTagConstant.php
+ */
+class PosTagConstant
+{
+    const A    = 'a';
+    const AD   = 'ad';
+    const AG   = 'ag';
+    const AN   = 'an';
+    const B    = 'b';
+    const C    = 'c';
+    const D    = 'd';
+    const DF   = 'df';
+    const DG   = 'dg';
+    const E    = 'e';
+    const ENG  = 'eng';
+    const F    = 'f';
+    const G    = 'g';
+    const H    = 'h';
+    const I    = 'i';
+    const J    = 'j';
+    const K    = 'k';
+    const L    = 'l';
+    const M    = 'm';
+    const MG   = 'mg';
+    const MQ   = 'mq';
+    const N    = 'n';
+    const NG   = 'ng';
+    const NR   = 'nr';
+    const NRFG = 'nrfg';
+    const NRT  = 'nrt';
+    const NS   = 'ns';
+    const NT   = 'nt';
+    const NZ   = 'nz';
+    const O    = 'o';
+    const P    = 'p';
+    const Q    = 'q';
+    const R    = 'r';
+    const RG   = 'rg';
+    const RR   = 'rr';
+    const RZ   = 'rz';
+    const S    = 's';
+    const T    = 't';
+    const TG   = 'tg';
+    const U    = 'u';
+    const UD   = 'ud';
+    const UG   = 'ug';
+    const UJ   = 'uj';
+    const UL   = 'ul';
+    const UV   = 'uv';
+    const UZ   = 'uz';
+    const V    = 'v';
+    const VD   = 'vd';
+    const VG   = 'vg';
+    const VI   = 'vi';
+    const VN   = 'vn';
+    const VQ   = 'vq';
+    const W    = 'w';
+    const X    = 'x';
+    const Y    = 'y';
+    const Z    = 'z';
+    const ZG   = 'zg';
+
+    const TAGS = [
+        self::A    => '形容词',
+        self::AD   => '副形词',
+        self::AG   => '形容词性语素',
+        self::AN   => '名形词',
+        self::B    => '区别词',
+        self::C    => '连词',
+        self::D    => '副词',
+        self::DF   => '副词*',
+        self::DG   => '副语素',
+        self::E    => '叹词',
+        self::ENG  => '外语',
+        self::F    => '方位词',
+        self::G    => '语素',
+        self::H    => '前接成分',
+        self::I    => '成语',
+        self::J    => '简称略语',
+        self::K    => '后接成分',
+        self::L    => '习用语',
+        self::M    => '数词',
+        self::MG   => '数语素',
+        self::MQ   => '数词*',
+        self::N    => '名词',
+        self::NG   => '名语素',
+        self::NR   => '人名',
+        self::NRFG => '名词*',
+        self::NRT  => '名词*',
+        self::NS   => '地名',
+        self::NT   => '机构团体',
+        self::NZ   => '其他专名',
+        self::O    => '拟声词',
+        self::P    => '介词',
+        self::Q    => '量词',
+        self::R    => '代词',
+        self::RG   => '代词语素',
+        self::RR   => '代词*',
+        self::RZ   => '代词*',
+        self::S    => '处所词',
+        self::T    => '时间词',
+        self::TG   => '时语素',
+        self::U    => '助词',
+        self::UD   => '助词*',
+        self::UG   => '助词*',
+        self::UJ   => '助词*',
+        self::UL   => '助词*',
+        self::UV   => '助词*',
+        self::UZ   => '助词*',
+        self::V    => '动词',
+        self::VD   => '副动词',
+        self::VG   => '动语素',
+        self::VI   => '动词*',
+        self::VN   => '名动词',
+        self::VQ   => '动词*',
+        self::W    => '标点符号',
+        self::X    => '非语素字',
+        self::Y    => '语气词',
+        self::Z    => '状态词',
+        self::ZG   => '状态词*',
+    ];
+}