+++ tests/Jieba/PossegTest.php
@@ -108,7 +108,9 @@ class PossegTest extends TestCase
             )
         );
 
-        $seg_list = (new Posseg(new Jieba()))->cut("这是一个伸手不见五指的黑夜。我叫孙悟空，我爱北京，我爱Python和C++。");
+        $seg_list = (new Posseg(new Jieba()))
+            ->cut("这是一个伸手不见五指的黑夜。我叫孙悟空，我爱北京，我爱Python和C++。")
+            ->toArray();
 
         $this->assertSame($case_array, $seg_list);
     }