+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.25.3 / TBD
+============
+
+  * New class \Jieba\PosTagConstant from text file /dict/pos_tag_readable.txt.
+  * Standardize data structure with new class \Jieba\Word and \Jieba\Words.
+
 0.25.2 / 2017-06-22
 ===================